refactor(ff-decode): move SeekMode and HardwareAccel into dedicated files
Extract SeekMode to seek.rs and HardwareAccel to hardware.rs following the
project convention of keeping lib.rs to pure re-exports. Move associated
unit tests to their respective files and rename them to the project naming
convention (feature_should_expected_result).

Resolves #708
